Fix: Ignore juju 3.3 tests
