+++ .github/workflows/integration_test.yaml
@@ -16,7 +16,7 @@ jobs:
       # charmcraft-channel: latest/edge
       charmcraft-repository: alithethird/charmcraft
       charmcraft-ref: feat/expressjs-extension
-      modules: '["test_charm.py", "test_cos.py", "test_database.py", "test_db_migration.py", "test_django.py", "test_django_integrations.py", "test_fastapi.py", "test_go.py", "test_integrations.py", "test_proxy.py", "test_workers.py", "test_tracing.py", "test_config.py", "test_smtp.py", "test_openfga.py", "test_minimal.py"]'
+      modules: '["test_charm.py", "test_database.py", "test_db_migration.py", "test_django.py", "test_django_integrations.py", "test_fastapi.py", "test_go.py", "test_integrations.py", "test_proxy.py", "test_workers.py", "test_tracing.py", "test_config.py", "test_smtp.py", "test_openfga.py", "test_minimal.py", "test_loki.py", "test_prometheus.py", "test_grafana.py"]'
 
       rockcraft-channel: latest/edge
       juju-channel: ${{ matrix.juju-version }}
