+++ tests/integration/integrations/test_openfga.py
@@ -19,29 +19,6 @@ logger = logging.getLogger(__name__)
     [
         ("flask_app", 8000),
         ("django_app", 8000),
-    ],
-)
-def test_openfga_integrations(
-    juju: jubilant.Juju,
-    openfga_app_fixture: App,
-    port,
-    request: pytest.FixtureRequest,
-    openfga_server_app: App,
-    postgresql_k8s: App,
-):
-    """
-    arrange: Build and deploy the charm. Integrate the charm with OpenFGA.
-    act: Send a read authorization models request from the charm.
-    assert: The request succeeds.
-    """
-    openfga_integrations(
-        juju, openfga_app_fixture, port, request, openfga_server_app, postgresql_k8s
-    )
-
-
-@pytest.mark.parametrize(
-    "openfga_app_fixture, port",
-    [
         ("fastapi_app", 8080),
         ("go_app", 8080),
     ],
@@ -60,19 +37,6 @@ def test_openfga_integrations_noble(
     act: Send a read authorization models request from the charm.
     assert: The request succeeds.
     """
-    openfga_integrations(
-        juju, openfga_app_fixture, port, request, openfga_server_app, postgresql_k8s
-    )
-
-
-def openfga_integrations(
-    juju: jubilant.Juju,
-    openfga_app_fixture: App,
-    port,
-    request: pytest.FixtureRequest,
-    openfga_server_app: App,
-    postgresql_k8s: App,
-):
     openfga_app = request.getfixturevalue(openfga_app_fixture)
     juju.wait(lambda status: status.apps[openfga_app.name].is_active)
     juju.wait(lambda status: status.apps[openfga_server_app.name].is_active)
